separated resolvers into separate files for readability and updated resolver imports to server.ts as well as the structure of resolvers in new instance of ApolloServer
