+++ server/resolvers/boardResolvers.js
@@ -1,0 +1,11 @@
+module.exports = {
+  Board: {
+    jobs: async (parent, args, { postgresDB }) => {
+      const boardId = parent._id;
+      const text = 'SELECT * FROM jobs WHERE boards_id=$1';
+      const params = [boardId];
+      const jobs = await postgresDB.query(text, params);
+      return jobs.rows;
+    },
+  },
+};
