+++ server/schema.js
@@ -1,4 +1,5 @@
 const { gql } = require('apollo-server');
+// const GraphQLDateTime = require('graphql-iso-date');
 
 const typeDefs = gql`
   scalar Date
